refactor(eslint-rules): custom workspace rule folder structure adapted
