--- tools/eslint-rules/README.md
@@ -1,3 +1,0 @@
-# eslint-rules
-
-This library was generated with [Nx](https://nx.dev).
