--- tools/eslint-rules/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@code-pushup-design-system/eslint-rules",
-  "version": "0.0.1",
-  "dependencies": {},
-  "private": true
-}
