+++ tools/eslint-rules/project.json
@@ -1,9 +1,14 @@
 {
   "name": "eslint-rules",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "tools/eslint-rules/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project eslint-rules --web",
-  "targets": {}
+  "sourceRoot": "tools/eslint-rules",
+  "targets": {
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "tools/eslint-rules/jest.config.ts"
+      }
+    }
+  }
 }
