+++ tools/eslint-rules/rules/rule-creator-factory.ts
